UPDATED: Controllers structure changed + reexport
